added many updated Files
Dont want to create tables, instead create a db and run the app.

+++ src/main/java/com/ford/respository/IAdminRepository.java
@@ -9,7 +9,6 @@ import java.util.Optional;
 @Repository
 public interface IAdminRepository extends JpaRepository<Admin, Integer> {
 
-
-    Admin findByAdminId(int adminId);
+    Admin findByAdminId(long adminId);
 }
