+++ src/main/java/com/ford/respository/ICourseRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ICourseRepository extends JpaRepository<Course, Integer> {
-    List<Course> findByAdmin_AdminId(int adminId);
+    List<Course> findByAdmin_AdminId(long adminId);
     Course findCourseByCourseId(int courseId);
+//    long countByAdminId(int adminId);
 
 }
