+++ src/main/java/com/ford/service/IAttachmentService.java
@@ -4,8 +4,8 @@ import com.ford.entity.Attachments;
 import org.springframework.http.ResponseEntity;
 
 public interface IAttachmentService {
-    ResponseEntity<String> createAttachments(int adminId, int courseId, Attachments attachments);
-//    ResponseEntity<String> deleteAttachments(int adminId, int courseId);
-//    ResponseEntity<String> getAttachments(int adminId, int courseId);
-    ResponseEntity<String> updateAttachments(int adminId,int courseId,int attachmentId,Attachments attachments);
+    ResponseEntity<String> createAttachments(long adminId, int courseId, Attachments attachments);
+//    ResponseEntity<String> deleteAttachments(long adminId, int courseId);
+//    ResponseEntity<String> getAttachments(long adminId, int courseId);
+    ResponseEntity<String> updateAttachments(long adminId,int courseId,int attachmentId,Attachments attachments);
 }
