+++ src/main/java/com/ford/service/ICourseService.java
@@ -9,7 +9,7 @@ import org.springframework.http.ResponseEntity;
 import java.util.List;
 
 public interface ICourseService{
-    ResponseEntity<List<Course>> findCoursesByAdmin(int adminId);
+    ResponseEntity<List<Course>> findCoursesByAdmin(long adminId);
     ResponseEntity<List<Course>> getAllCourses();
     ResponseEntity<Course> findCourseByCourseId(int courseId);
 }
